requirements.txt
python-dotenv==1.0.1
simple-salesforce==1.12.6
pandas==2.2.2
pyarrow==16.1.0
boto3==1.34.131
requests==2.31.0